[Fixed] Copy extradata

+++ fpp/core/Utils.hpp
@@ -29,7 +29,7 @@ namespace fpp {
         static std::string  to_string(MediaType type);
         static std::string  pts_to_string(int64_t pts);
         static std::string  time_to_string(int64_t time_stamp, AVRational time_base);
-        static std::string  channel_layout_to_string(int64_t nb_channels, uint64_t channel_layout);
+        static std::string  channel_layout_to_string(int nb_channels, uint64_t channel_layout);
 
         static MediaType    to_media_type(AVMediaType type);
         static AVMediaType  from_media_type(MediaType type);
